// src/common/enums/service-request-status.enum.ts
export enum ServiceRequestStatus {
  RECEIVED = 'received',
  CANCELLED = 'cancelled',
  REJECTED = 'rejected',
  APPROVED = 'approved',
  UNPROCESSABLE = 'unprocessable',
  ON_PROGRESS = 'on_progress',
  PARTIAL_COMPLETED = 'partial_completed',
  FAILED = 'failed',
  COMPLETED = 'completed',
  CLOSED = 'closed',
}